extract utility function into util crate
