+++ node/network/collator-protocol/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2018"
 futures = "0.3.5"
 log = "0.4.11"
 derive_more = "0.99.9"
-pin-project = "0.4.22"
 
 codec = { package="parity-scale-codec", version = "1.3.4", features = ["std"]  }
